Switch to Gson instead of jackson for reading json strings
