+++ src/main/java/com/zendesk/Main.java
@@ -1,4 +1,4 @@
-package zendesk;
+package com.zendesk;
 
 public class Main {
