+++ src/main/java/com/zendesk/util/Constants.java
@@ -1,0 +1,7 @@
+package com.zendesk.util;
+
+public class Constants {
+
+  public static final String DATE_FORMAT = "yyyy-MM-dd'T'HH:mm:ss z";
+
+}
